All the changes for #7
All messages are now handled by the class mapped to the destination
system, instead of the source. This makes a lot more sense.
