+++ src/net/littlebigisland/droidibus/ibus/DeviceAddressEnum.java
@@ -10,10 +10,10 @@ public enum DeviceAddressEnum {
 	CDChanger((byte) 0x18),
 	RadioControlledClock((byte) 0x28),
 	CheckControlModule((byte) 0x30),
-	GraphicsNavigationDriver((byte) 0x3B),
+	GFXNavigationDriver((byte) 0x3B),
 	Diagnostic((byte) 0x3F),
 	RemoteControlCentralLocking((byte) 0x40),
-	GraphicsDriverRearScreen((byte) 0x43),
+	GFXDriverRearScreen((byte) 0x43),
 	Immobiliser((byte) 0x44),
 	CentralInformationDisplay((byte) 0x46),
 	MultiFunctionSteeringWheel((byte) 0x50),
@@ -33,7 +33,7 @@ public enum DeviceAddressEnum {
 	AirBagModule((byte) 0xA4),
 	SpeedRecognitionSystem((byte) 0xB0),
 	NavigationJapan((byte) 0xBB),
-	GlobalBroadcastAddress((byte) 0xBF),
+	GlobalBroadcast((byte) 0xBF),
 	MultiInfoDisplay((byte) 0xC0),
 	Telephone((byte) 0xC8),
 	Assist((byte) 0xCA),
@@ -43,7 +43,7 @@ public enum DeviceAddressEnum {
 	FrontDisplay((byte) 0xE7),
 	RainLightSensor((byte) 0xE8),
 	Television((byte) 0xED),
-	OnBoardMonitor((byte) 0xF0),
+	BoardMonitor((byte) 0xF0),
 	CSU((byte) 0xF5),
 	Broadcast((byte) 0xFF),
 	Unset((byte) 0x100),
