+++ src/net/littlebigisland/droidibus/ibus/IBusCommandsEnum.java
@@ -4,51 +4,51 @@ package net.littlebigisland.droidibus.ibus;
  * This should be better explained. Basically, we hold an enum with two values,
  * the source system and the method name that actually creates the data for the method
  * we want. Of course the Source system tells you which class you'll find these methods in.
- * e.x. OnBoardMonitor == BoardMonitorSystemCommand
+ * e.x. BoardMonitor == BoardMonitorSystemCommand
  */
 public enum IBusCommandsEnum {
 	// BoardMonitor to IKE Commands
-	BMToIKEGetIgnitionStatus(DeviceAddressEnum.OnBoardMonitor, "getIgnitionStatus"),
-	BMToIKEGetTime(DeviceAddressEnum.OnBoardMonitor, "getTime"),
-	BMToIKEGetDate(DeviceAddressEnum.OnBoardMonitor, "getDate"),
-	BMToIKEGetOutdoorTemp(DeviceAddressEnum.OnBoardMonitor, "getOutdoorTemp"),
-	BMToIKEGetFuel1(DeviceAddressEnum.OnBoardMonitor, "getFuel1"),
-	BMToIKEGetFuel2(DeviceAddressEnum.OnBoardMonitor, "getFuel2"),
-	BMToIKEGetRange(DeviceAddressEnum.OnBoardMonitor, "getRange"),
-	BMToIKEGetAvgSpeed(DeviceAddressEnum.OnBoardMonitor, "getAvgSpeed"),
-	BMToIKEResetFuel1(DeviceAddressEnum.OnBoardMonitor, "resetFuel1"),
-	BMToIKEResetFuel2(DeviceAddressEnum.OnBoardMonitor, "resetFuel2"),
-	BMToIKEResetAvgSpeed(DeviceAddressEnum.OnBoardMonitor, "resetAvgSpeed"),
-	BMToIKESetTime(DeviceAddressEnum.OnBoardMonitor, "setTime"),
-	BMToIKESetDate(DeviceAddressEnum.OnBoardMonitor, "setDate"),
-	BMToIKESetUnits(DeviceAddressEnum.OnBoardMonitor, "setUnits"),
+	BMToIKEGetIgnitionStatus(DeviceAddressEnum.BoardMonitor, "getIgnitionStatus"),
+	BMToIKEGetTime(DeviceAddressEnum.BoardMonitor, "getTime"),
+	BMToIKEGetDate(DeviceAddressEnum.BoardMonitor, "getDate"),
+	BMToIKEGetOutdoorTemp(DeviceAddressEnum.BoardMonitor, "getOutdoorTemp"),
+	BMToIKEGetFuel1(DeviceAddressEnum.BoardMonitor, "getFuel1"),
+	BMToIKEGetFuel2(DeviceAddressEnum.BoardMonitor, "getFuel2"),
+	BMToIKEGetRange(DeviceAddressEnum.BoardMonitor, "getRange"),
+	BMToIKEGetAvgSpeed(DeviceAddressEnum.BoardMonitor, "getAvgSpeed"),
+	BMToIKEResetFuel1(DeviceAddressEnum.BoardMonitor, "resetFuel1"),
+	BMToIKEResetFuel2(DeviceAddressEnum.BoardMonitor, "resetFuel2"),
+	BMToIKEResetAvgSpeed(DeviceAddressEnum.BoardMonitor, "resetAvgSpeed"),
+	BMToIKESetTime(DeviceAddressEnum.BoardMonitor, "setTime"),
+	BMToIKESetDate(DeviceAddressEnum.BoardMonitor, "setDate"),
+	BMToIKESetUnits(DeviceAddressEnum.BoardMonitor, "setUnits"),
 	// BoardMonitor to Radio Commands
-	BMToRadioGetStatus(DeviceAddressEnum.OnBoardMonitor, "getRadioStatus"),
-	BMToRadioCDStatus(DeviceAddressEnum.OnBoardMonitor, "sendCDPlayerMessage"),
-	BMToRadioPwrPress(DeviceAddressEnum.OnBoardMonitor, "sendRadioPwrPress"),
-	BMToRadioPwrRelease(DeviceAddressEnum.OnBoardMonitor, "sendRadioPwrRelease"),
-	BMToRadioTonePress(DeviceAddressEnum.OnBoardMonitor, "sendTonePress"),
-	BMToRadioToneRelease(DeviceAddressEnum.OnBoardMonitor, "sendToneRelease"),
-	BMToRadioModePress(DeviceAddressEnum.OnBoardMonitor, "sendModePress"),
-	BMToRadioModeRelease(DeviceAddressEnum.OnBoardMonitor, "sendModeRelease"),
-	BMToRadioVolumeUp(DeviceAddressEnum.OnBoardMonitor, "sendVolumeUp"),
-	BMToRadioVolumeDown(DeviceAddressEnum.OnBoardMonitor, "sendVolumeDown"),
-	BMToRadioTuneFwdPress(DeviceAddressEnum.OnBoardMonitor, "sendSeekFwdPress"),
-	BMToRadioTuneFwdRelease(DeviceAddressEnum.OnBoardMonitor, "sendSeekFwdRelease"),
-	BMToRadioTuneRevPress(DeviceAddressEnum.OnBoardMonitor, "sendSeekRevPress"),
-	BMToRadioTuneRevRelease(DeviceAddressEnum.OnBoardMonitor, "sendSeekRevRelease"),
-	BMToRadioFMPress(DeviceAddressEnum.OnBoardMonitor, "sendFMPress"),
-	BMToRadioFMRelease(DeviceAddressEnum.OnBoardMonitor, "sendFMRelease"),
-	BMToRadioAMPress(DeviceAddressEnum.OnBoardMonitor, "sendAMPress"),
-	BMToRadioAMRelease(DeviceAddressEnum.OnBoardMonitor, "sendAMRelease"),
-	BMToRadioInfoPress(DeviceAddressEnum.OnBoardMonitor, "sendInfoPress"),
-	BMToRadioInfoRelease(DeviceAddressEnum.OnBoardMonitor, "sendInfoRelease"),
+	BMToRadioGetStatus(DeviceAddressEnum.BoardMonitor, "getRadioStatus"),
+	BMToRadioCDStatus(DeviceAddressEnum.BoardMonitor, "sendCDPlayerMessage"),
+	BMToRadioPwrPress(DeviceAddressEnum.BoardMonitor, "sendRadioPwrPress"),
+	BMToRadioPwrRelease(DeviceAddressEnum.BoardMonitor, "sendRadioPwrRelease"),
+	BMToRadioTonePress(DeviceAddressEnum.BoardMonitor, "sendTonePress"),
+	BMToRadioToneRelease(DeviceAddressEnum.BoardMonitor, "sendToneRelease"),
+	BMToRadioModePress(DeviceAddressEnum.BoardMonitor, "sendModePress"),
+	BMToRadioModeRelease(DeviceAddressEnum.BoardMonitor, "sendModeRelease"),
+	BMToRadioVolumeUp(DeviceAddressEnum.BoardMonitor, "sendVolumeUp"),
+	BMToRadioVolumeDown(DeviceAddressEnum.BoardMonitor, "sendVolumeDown"),
+	BMToRadioTuneFwdPress(DeviceAddressEnum.BoardMonitor, "sendSeekFwdPress"),
+	BMToRadioTuneFwdRelease(DeviceAddressEnum.BoardMonitor, "sendSeekFwdRelease"),
+	BMToRadioTuneRevPress(DeviceAddressEnum.BoardMonitor, "sendSeekRevPress"),
+	BMToRadioTuneRevRelease(DeviceAddressEnum.BoardMonitor, "sendSeekRevRelease"),
+	BMToRadioFMPress(DeviceAddressEnum.BoardMonitor, "sendFMPress"),
+	BMToRadioFMRelease(DeviceAddressEnum.BoardMonitor, "sendFMRelease"),
+	BMToRadioAMPress(DeviceAddressEnum.BoardMonitor, "sendAMPress"),
+	BMToRadioAMRelease(DeviceAddressEnum.BoardMonitor, "sendAMRelease"),
+	BMToRadioInfoPress(DeviceAddressEnum.BoardMonitor, "sendInfoPress"),
+	BMToRadioInfoRelease(DeviceAddressEnum.BoardMonitor, "sendInfoRelease"),
 	// BoardMonitor to Light Control Module Commands
-	BMToLCMGetDimmerStatus(DeviceAddressEnum.OnBoardMonitor, "getLightDimmerStatus"),
+	BMToLCMGetDimmerStatus(DeviceAddressEnum.BoardMonitor, "getLightDimmerStatus"),
 	// BoardMonitor to General Module Commands
-	BMToGMGetDoorStatus(DeviceAddressEnum.OnBoardMonitor, "getDoorsRequest"),
+	BMToGMGetDoorStatus(DeviceAddressEnum.BoardMonitor, "getDoorsRequest"),
 	// BoardMonitor to Global Broadcast Address Commands
-	BMToGlobalBroadcastAliveMessage(DeviceAddressEnum.OnBoardMonitor, "sendAliveMessage"),
+	BMToGlobalBroadcastAliveMessage(DeviceAddressEnum.BoardMonitor, "sendAliveMessage"),
 	// Steering Wheel to Radio Commands
 	SWToRadioVolumeUp(DeviceAddressEnum.MultiFunctionSteeringWheel, "sendVolumeUp"),
 	SWToRadioVolumeDown(DeviceAddressEnum.MultiFunctionSteeringWheel, "sendVolumeDown"),
